backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
fpdf2
websockets
aiohttp